make address manager public
