+++ integration/addr/addr_suite_test.go
@@ -1,0 +1,14 @@
+package addr_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestAddr(t *testing.T) {
+	t.Parallel()
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Addr Suite")
+}
